Modified entry_point to run hub/node combination
In my attempts to have another container utilize this as a Remote
selenium driver, I had issues with only being allowed to make requests
from within the container. By running both a hub and node you will be
able to have other containers use this container for driving a browser.
Therefore your test container does not have to derive from this image.

For example if you build a NodeJs, Python, or Java application you can
run your server and tests from your application's container with a
browser from this one.

However, if you still wish to derive from this container and make calls
to within itself you are also free to do so.