Made it possible to build customers version of these docker images.
I tried to build private copies of these images and manually set
NAME and VERSION when launching make. Unfortunately it turned out
that the NAME did not propagate so for example NodeFirefox wanted to
build on selenium/node-base:MY_VERSION which failed since that didn't
exist. This patch propagates NAME to the generate.sh-scripts so the
images build on each other.