Fix for issue with not being able to pass environment variables into node containers at runtime
When changes were introduced to generate default config during build
time with #502 and then with #529, it broke the ability to pass in
environment variables to the node containers at runtime. This fix aims
to provide the ability to do both i.e. continue to pass environment
variables at runtime and generate a default config at build time for
the node.