Fix Chrome beta support for NodeChrome
The Dockerfile says Chrome beta is supported:
>  can specify versions by CHROME_VERSION;
>  e.g. google-chrome-stable=53.0.2785.101-1
>       google-chrome-beta=53.0.2785.92-1

But when actually building an image with e.g.
`--build-arg CHROME_VERSION=google-chrome-beta=62.0.3202.29-1`, then a couple
of problems appear. Firstly, the version field in `/opt/selenium/config.json`
will not be populated, because `/opt/google/chrome/chrome` does not exist.
Secondly and even more importantly, the node simply won't work. I assume it's
because `chrome_launcher.sh` tries to execute the nonexistent
`/opt/google/chrome/chrome`, but I haven't confirmed for sure.

This change fixes both of those issues by actually wrapping the existing
executable, instead of replacing it. I have locally tested it by with latest
versions (not specifying any build arguments) and with the latest Chrome beta
version. Both work.

As can be seen from current [Chromium source][1], the wrapping code that was
used in chrome_launcher.sh has changed considerably. This change avoids
duplicating that code here and so makes it impossible for this to get out of
date. In addition, this will work with different Chrome versions which might
have different code in the wrapper.

[1]: https://cs.chromium.org/chromium/src/chrome/installer/linux/common/wrapper