[grid][chart] Allows users to turn of Deployment creation for Nodes
This change introduces a new value called `deploymentEnabled` for each of the node browsers,
if the value is set to `true` Deployments will be created for each of the respective nodes however if `false` no Deployments will be created.

This will allow users to turn of creation of the default deployments and to use other resource types instead such as jobs.

Fixes #1708