Do not print secrets such as registration secret and router password to console on selenium grid hub and router startup
When using the start-selenium-grid-hub.sh or start-selenium-grid-router.sh script the router password and registration secret got printed to the console.
For testing usage this is (semi) fine. But when using Selenium Docker Images in a production environment, printing secrets to the console is a security issue.
Therefore, all secrets printed to the console are replaced with '***' in this change.