Docker: Respect SE_SCREEN_{WIDTH,HEIGHT} variables in ffmpeg recording
Currently, these variables are always overridden in the call the
wait_for_display(). Before the upgrade to ffmpeg-7.0.2 (PR #2374), the
SE_SCREEN_WIDTH and SE_SCREEN_HEIGHT variables were respected in the
case where no video uploading was done because wait_for_display() was
never called. After the upgrade to 7.0.2, it the function is always
called and overrides the VIDEO_SIZE variable.

Now, if both screen width and height are specified, the VIDEO_SIZE is
not overridden and users can control the recording size again.